modify port id from uint32 to uint64

+++ ap_verifier.h
@@ -21,7 +21,7 @@ public:
     uint32_t ap_size;
 
     // topology
-    std::map< uint32_t, std::vector<uint32_t>* > topology;
+    std::map< uint64_t, std::vector<uint64_t>* > topology;
 
     // router map: from router_id to predicate list;
     std::map< uint32_t, Router* > id_to_router;
@@ -33,14 +33,14 @@ public:
     std::vector< bdd >* ap_bdd_list;
 
     // map from port_id to router_id
-    std::map< uint32_t, uint32_t > inport_to_router;
+    std::map< uint64_t, uint32_t > inport_to_router;
 
     APVerifier(int length, AP_TYPE type);
     int get_length() { return this->hdr_len; }
 
     ~APVerifier();
 
-    void add_link(uint32_t from_port, uint32_t to_port);
+    void add_link(uint64_t from_port, uint64_t to_port);
 
     void print_topology();
 
@@ -52,16 +52,16 @@ public:
 
     void convert_router_to_ap();
 
-    void query_reachability(uint32_t from_port, uint32_t to_port);
+    void query_reachability(uint64_t from_port, uint64_t to_port);
 
-    void propagate_bdd(bdd packet_header, std::list< uint32_t > passed_port, uint32_t from_port,
-                       uint32_t dst_port);
+    void propagate_bdd(bdd packet_header, std::list< uint64_t > passed_port, uint64_t from_port,
+                       uint64_t dst_port);
 
-    void propagate_vec(std::vector< bool > packet_header, std::list< uint32_t > passed_port, uint32_t from_port,
-                   uint32_t dst_port);
+    void propagate_vec(std::vector< bool > packet_header, std::list< uint64_t > passed_port, uint64_t from_port,
+                   uint64_t dst_port);
 
-    void propagate_bset(std::bitset<BITSETLEN> packet_header, std::list< uint32_t > passed_port, uint32_t from_port,
-                       uint32_t dst_port);
+    void propagate_bset(std::bitset<BITSETLEN> packet_header, std::list< uint64_t > passed_port, uint64_t from_port,
+                       uint64_t dst_port);
 };
 
 bdd match2bdd(string match, int length);
