+++ ap_verifier_utils.h
@@ -26,7 +26,7 @@ enum AP_TYPE {
 
 struct PACKED List_t {
     uint32_t size;
-    uint32_t *list;
+    uint64_t *list;
     bool shared;
 };
 
@@ -36,11 +36,11 @@ extern log4cxx::LoggerPtr rlogger;
 
 extern bool show_detail;
 
-List_t make_sorted_list_from_array (uint32_t count, uint32_t elems[]);
+List_t make_sorted_list_from_array (uint32_t count, uint64_t elems[]);
 List_t val_to_list(const Json::Value &val);
 std::string list_to_string(List_t p);
 void print_bool_vector(std::vector< bool > packet_header);
-void print_passed_port(std::list< uint32_t > passed_port);
+void print_passed_port(std::list< uint64_t > passed_port);
 
 
 #endif //AP_VERIFIER_AP_VERIFIER_UTILS_H
